Resolve require cycle issue from rn-fetch-blob
rn-fetch-blob is no longer maintained and a [require cycle issue](https://github.com/joltup/rn-fetch-blob/issues/183) constantly shows warnings in the latest version of RN when we start the app. This updates the package to use [react-native-blob-util](https://github.com/RonRadtke/react-native-blob-util) instead which is a maintained fork of the original package.